CScriptObject: Make more extensive use of CInstanceID
We had parts of the core API that were just using raw values when this
class already exists, so we can flesh this out a little more to make
better use of it.

While we're at it, we can drop all the implicit conversion facilities to
make it very explicit when we're constructing these instance IDs, making
it easier to visually track.

+++ src/Core/Resource/Script/CScriptTemplate.cpp
@@ -179,7 +179,7 @@ int32 CScriptTemplate::CheckVolumeConditions(CScriptObject *pObj, bool LogErrors
         if (LogErrors)
         {
             NLog::Error("{} instance {:08X} has unexpected volume shape value of 0x{:X}", *pObj->Template()->Name(),
-                        static_cast<uint32>(pObj->InstanceID()), Val);
+                        pObj->InstanceID(), Val);
         }
     }
